+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/SwirldsPlatform.java
@@ -109,7 +109,7 @@ import com.swirlds.platform.event.validation.AncientValidator;
 import com.swirlds.platform.event.validation.EventDeduplication;
 import com.swirlds.platform.event.validation.GossipEventValidator;
 import com.swirlds.platform.event.validation.GossipEventValidators;
-import com.swirlds.platform.event.validation.LegacyEventValidator;
+import com.swirlds.platform.event.validation.EventValidator;
 import com.swirlds.platform.event.validation.SignatureValidator;
 import com.swirlds.platform.event.validation.StaticValidators;
 import com.swirlds.platform.event.validation.TransactionSizeValidator;
@@ -215,7 +215,7 @@ public class SwirldsPlatform implements Platform {
     /**
      * Validates events and passes valid events further down the pipeline.
      */
-    private final LegacyEventValidator eventValidator;
+    private final EventValidator eventValidator;
     /** Contains all validators for events */
     private final GossipEventValidators eventValidators;
 
@@ -612,7 +612,7 @@ public class SwirldsPlatform implements Platform {
         }
 
         eventValidators = new GossipEventValidators(validators);
-        eventValidator = new LegacyEventValidator(
+        eventValidator = new EventValidator(
                 eventValidators, eventIntake::addUnlinkedEvent, eventIntakePhaseTimer, intakeEventCounter);
 
         intakeQueue = components.add(new QueueThreadConfiguration<GossipEvent>(threadManager)
