+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/preconsensus/PreconsensusEventReplayPipeline.java
@@ -24,7 +24,7 @@ import com.swirlds.common.threading.framework.config.QueueThreadConfiguration;
 import com.swirlds.common.threading.framework.config.ThreadConfiguration;
 import com.swirlds.common.threading.manager.ThreadManager;
 import com.swirlds.platform.event.GossipEvent;
-import com.swirlds.platform.event.validation.LegacyEventValidator;
+import com.swirlds.platform.event.validation.EventValidator;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -62,7 +62,7 @@ public class PreconsensusEventReplayPipeline {
 
     private final PlatformContext platformContext;
     private final IOIterator<GossipEvent> unhashedEventIterator;
-    private final LegacyEventValidator eventValidator;
+    private final EventValidator eventValidator;
 
     /**
      * Create a new event replay pipeline.
@@ -76,7 +76,7 @@ public class PreconsensusEventReplayPipeline {
             @NonNull final PlatformContext platformContext,
             @NonNull final ThreadManager threadManager,
             @NonNull final IOIterator<GossipEvent> unhashedEventIterator,
-            @NonNull final LegacyEventValidator eventValidator) {
+            @NonNull final EventValidator eventValidator) {
 
         this.platformContext = Objects.requireNonNull(platformContext);
         Objects.requireNonNull(threadManager);
