+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/event/validation/EventValidatorTests.java
@@ -37,7 +37,7 @@ import com.swirlds.platform.event.GossipEvent;
 import com.swirlds.platform.event.validation.EventDeduplication;
 import com.swirlds.platform.event.validation.GossipEventValidator;
 import com.swirlds.platform.event.validation.GossipEventValidators;
-import com.swirlds.platform.event.validation.LegacyEventValidator;
+import com.swirlds.platform.event.validation.EventValidator;
 import com.swirlds.platform.event.validation.StaticValidators;
 import com.swirlds.platform.event.validation.TransactionSizeValidator;
 import com.swirlds.platform.gossip.IntakeEventCounter;
@@ -112,7 +112,7 @@ class EventValidatorTests {
     void eventValidator() {
         final Set<GossipEvent> intakeEvents = new HashSet<>();
         final AtomicBoolean isValid = new AtomicBoolean(true);
-        final LegacyEventValidator eventValidator = new LegacyEventValidator(
+        final EventValidator eventValidator = new EventValidator(
                 (e) -> isValid.get(), intakeEvents::add, mock(PhaseTimer.class), mock(IntakeEventCounter.class));
 
         final GossipEvent validEvent = GossipEventBuilder.builder().buildEvent();
